pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Alexsandr250897 <132817845+Alexsandr250897@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
postgres = "^4.0"
djangorestframework = "^3.15.2"
pytest = "^8.3.3"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
